Fix non-determistic incorrect output with respect to nsPrefixes
This change fixes the test failures on mbg/wai-saml2#52 . It turns out that the use of `unsafeAsCString` causes `nsPrefixes` to be effectively empty most of the time (yes, it is __non-deterministic__!)

I confirmed that the tests above with this patch pass on lts-19.33 or older.

I suggest to release the fix in an urgent manner.